refactor: improve error handling in messageService and websocketService
